Tutorial: Large number of improvements
Make a large number of changes to the tutorial
so that it now matches the current mmj2 and set.mm.
This also improves the overall flow.

This swaps pages 402 and 403; I think this order
is much easier to understand.

It also adds a new Page503, which explains
"File/Get Proof" and hints a little bit at the
set.mm label naming conventions.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>