Increase maximum Java memory to 2 Gibibytes
Increase the maximum Java memory to 2 Gibibytes (2G)
when running using the provided Windows batch files.
Without this change, mmj2 can fail on Windows
due to a lack of memory when processing set.mm.

The file set.mm keeps getting bigger, and mmj2
includes an Early parser (which uses more memory),
so the old maximum of 512 Mebibytes (512M)
can lead to failure on startup when processing set.mm.

This commit increases the maximum to 2 Gibibytes.
This seems to be enough to handle set.mm.
This maximum is a strain for 32-bit systems, but such
systems are disappearing, and by the time we might need
to re-raise it I don't expect 32-bit systems to be in
use for this purposes.

This commit does *not* change the starting memory size,
just the maximum, because a user might choose a database
other than set.mm which doesn't require a lot of memory.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>