Increase the maximum Java memory to 1280 Mebibytes
Increase the maximum Java memory to 1280 Mebibytes (MiB)
when using the provided Windows batch files.
Without this change, mmj2 is likely to fail on Windows
when processing big databases (especially set.mm)
due to a lack of memory.

This maximum is carefully selected to work on 32-bit Windows
as well as 64-bit Windows, so it should mean that mmj2
will "just work" on any plausible Windows system.

The underlying cause of the problem is that
the file set.mm keeps getting bigger and that mmj2
includes an Earley parser (which uses more memory).
As a result, the old maximum of 512 Mebibytes (512MiB)
is likely to lead to startup failure when processing set.mm.

This commit increases the maximum to 1280 Mebibytes
(1024MiB + 256MiB). Currently 1024MiB is enough to handle
set.mm; adding another 256MiB provides more
room for growth without issue.
This number is intentionally under 1500MiB, which is
around the most that 32-bit Windows systems can handle
(the actual number varies, this commit is conservative).
64-bit systems can support a much larger maximum, but
there is no expectation that a larger number will be needed
for many years, and by that time we think it will be reasonable
to require the use of 64-bit system for proof development
even for people with limited resources.

This commit does not change the starting memory size,
just the maximum, because a user might choose a database
other than set.mm which doesn't require a lot of memory.

This commit also converts the .bat files (but *only* the .bat
files) to MS-DOS text format (CRLF for end-of-line). These files are
*only* executed by Windows systems, and some Windows programs
(particularly most versions of Notepad) can only handle
MS-DOS text format well. Switching *only* those files should make
it easier for Windows users to get programs started.
Everything else continues to use the POSIX standard text format
(where lines end in only LF); once mmj2 is started, the mmj2
text editor can handle POSIX text format properly.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>