mmj2: Rewrite script so it works "normally"
Create a new mmj2 entry script that works "normally", that is,
it accept pathnames to .mm or .mmp files, has parameters for
the database or run command file, and so on.

Usage:
mmj2 [ -f RUNCOMMANDS ] [ -d METAMATH_DATABASE_FILE ] [ FILE ]
- If FILE is a .mmp file, open that .mmp file when starting out.
- If FiLE is a .mm file, open that database and use a "default" .mmp file.
If METAMATH_DATABASE_FILE isn't provided, it makes an effort to find it,
and records its location in ~/.config/mmj2/default_database for next time.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>