Move contact info for Mel O'Cat to a different file
We are very grateful to Mel O'Cat for his work on mmj2,
but this is really a fork of his earlier work, and we don't
want people to think this is exactly the same program he wrote.
So instead, move the contact info for Mel O'Cat into a separate file,
and ensure that the README points to it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>