mmj2.bat: If named database not found, search for it
This modifies mmj2.bat so it has the same capability as the
mmj2 shell command: if you give it a datbase name, but the
named database does not exist, search for it.
E.g., if you say `mmj2.bat iset.mm`, and `iset.mm` is not
present in the current directory, it will be searched for
(by default using the built-in list of directories).

We also document this capability, since it works in shell and batch.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>